refactor: 커스텀 RedisOperator 인터페이스 추가
