+++ src/main/java/com/koddy/server/global/redis/StringRedisOperator.java
@@ -1,4 +1,4 @@
-package com.koddy.server.global.utils;
+package com.koddy.server.global.redis;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.StringRedisTemplate;
@@ -8,21 +8,25 @@ import java.time.Duration;
 
 @Component
 @RequiredArgsConstructor
-public class DefaultRedisStringOperator {
+public class StringRedisOperator implements RedisOperator<String, String> {
     private final StringRedisTemplate redisTemplate;
 
+    @Override
     public void save(final String key, final String value, final Duration duration) {
         redisTemplate.opsForValue().set(key, value, duration);
     }
 
+    @Override
     public String get(final String key) {
         return redisTemplate.opsForValue().get(key);
     }
 
+    @Override
     public boolean contains(final String key) {
         return redisTemplate.opsForValue().get(key) != null;
     }
 
+    @Override
     public void delete(final String key) {
         redisTemplate.delete(key);
     }
